Add files via upload
Second version of the function that evolves data-splitting solutions to build a decision tree using a genetic algorithm. I reorganized some parts of the function and added an improvement that reduces execution time in the end.  It remains basically the same, but now it includes a limit beyond which the GA stops running generations if no performance improvement is found between consecutive generations.

NOTE THAT THE FUNCTIONS ARE EXACTLY THE SAME, I ONLY ADDED QUALITY-OF-LIFE CHANGES.